+++ allink_core/allink_base/admin/forms.py
@@ -25,7 +25,7 @@ class AllinkBaseAdminForm(TranslatableModelForm):
                     is_stacked=True
                 ),
                 required=True,
-                queryset=AllinkCategory.relevant_categories(self.instance._meta.model_name)
+                queryset=self.instance.get_relevant_categories()
             )
             self.fields['category_navigation'] = forms.ModelMultipleChoiceField(
                 label=_(u'Categories for Navigation'),
@@ -36,7 +36,7 @@ class AllinkBaseAdminForm(TranslatableModelForm):
                 help_text=_(
                     u'You can explicitly define the categories for the category navigation here. This will override the automatically set of categories. (From "Filter & Ordering" but not from the "Manual entries")'),
                 required=False,
-                queryset=AllinkCategory.relevant_categories(self.instance._meta.model_name)
+                queryset=self.instance.get_relevant_categories()
             )
 
 
@@ -57,7 +57,7 @@ class AllinkBaseAppContentPluginForm(forms.ModelForm):
                     is_stacked=True
                 ),
                 required=False,
-                queryset=AllinkCategory.relevant_categories(self.instance.data_model._meta.model_name)
+                queryset=self.instance.data_model.get_relevant_categories()
             )
             self.fields['categories_and'] = forms.ModelMultipleChoiceField(
                 label=_(u'Categories (AND operator)'),
@@ -67,7 +67,7 @@ class AllinkBaseAppContentPluginForm(forms.ModelForm):
                 ),
                 help_text=_(u'Use this field if you want to further restrict your result set. This option allows you to create a conjunction between the first set of categories in field "Categories" and the ones specified here.'),
                 required=False,
-                queryset=AllinkCategory.relevant_categories(self.instance.data_model._meta.model_name)
+                queryset=self.instance.data_model.get_relevant_categories()
             )
             self.fields['category_navigation'] = forms.ModelMultipleChoiceField(
                 label=_(u'Categories for Navigation'),
@@ -78,7 +78,7 @@ class AllinkBaseAppContentPluginForm(forms.ModelForm):
                 help_text=_(
                     u'You can explicitly define the categories for the category navigation here. This will override the automatically set of categories. (From "Filter & Ordering" but not from the "Manual entries")'),
                 required=False,
-                queryset=AllinkCategory.relevant_categories(self.instance.data_model._meta.model_name)
+                queryset=self.instance.data_model.get_relevant_categories()
             )
         self.fields['filter_fields'] = forms.TypedMultipleChoiceField(
             label=_(u'Filter Fields'),